+++ plugins/web_adapter/CHANGELOG.md
@@ -1,6 +1,17 @@
+# CHANGELOG
+
 ## Unreleased
 
 *None.*
+
+## 2.1.1
+
+- Move all source Dart files to `*_impl.dart` to avoid naming collision.
+  This is a workaround of https://github.com/dart-lang/sdk/issues/56498.
+
+## 2.1.0
+
+- Support `FileAccessMode` in `Dio.download` and `Dio.downloadUri` to change download file opening mode.
 
 ## 2.0.0